refactor: migrated connect wallet modal to connect wallet button component
